@@ -2,41 +2,9 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-kapt'
 apply plugin: 'kotlin-android-extensions'
-apply plugin: 'checkstyle'
-
-checkstyle {
-    toolVersion = '7.6.1'
-}
-
-preBuild.dependsOn 'checkstyle'
-check.dependsOn 'checkstyle'
-
-// auto increment version stuff
-import com.moldedbits.android.MBIncrementVersionTask
-
-task incrementVersionTask(type: MBIncrementVersionTask)
-tasks.whenTaskAdded { task ->
-    if (task.name == 'generateReleaseBuildConfig') {
-        task.dependsOn 'incrementVersionTask'
-    }
-}
-
-task checkstyle(type: Checkstyle) {
-    configFile file("${project.rootDir}/config/quality/checkstyle.xml")
-    source 'src'
-    include '**/*.java'
-    exclude '**/gen/**'
-    exclude '**/test/java/com/moldedbits/android/**'
-
-    classpath = files()
-
-    // set to false if you want to consider warning as error
-    ignoreFailures = false
-}
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 29
 
     final API_URL_DEV = "https://www.google.com/"
     final API_URL_PROD = "https://www.google.com/"
