@@ -1,7 +1,7 @@
 package com.moldedbits.android
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.view.MenuItem
 import android.widget.FrameLayout
 import com.moldedbits.android.dialogs.LoadingDialog
