@@ -1,7 +1,7 @@
 package com.moldedbits.android
 
 import android.os.Bundle
-import android.support.v7.widget.Toolbar
+import androidx.appcompat.widget.Toolbar
 import android.widget.FrameLayout
 
 abstract class BaseToolBarActivity : BaseActivity() {
